+++ NEWS.md
@@ -1,6 +1,6 @@
 # Signac 0.1.3
 
-* New `Harmonize` function to run Harmony on scATAC-seq data  
+* New `GetIntersectingFeatures` function to find overlapping peaks between objects  
 * New `MergeWithRegions` function to perform region-aware Seurat object merging  
 
 # Signac 0.1.2